Solução: manter comp. Button no App.tsx
