+++ src/components/Content.tsx
@@ -1,3 +1,52 @@
-export function Content() {
-  // Complete aqui
+import { useEffect, useState } from "react";
+import { MovieCard } from "./MovieCard";
+import { api } from '../services/api';
+
+interface MovieProps {
+  imdbID: string;
+  Title: string;
+  Poster: string;
+  Ratings: Array<{
+    Source: string;
+    Value: string;
+  }>;
+  Runtime: string;
+}
+interface GenreResponseProps {
+  id: number;
+  name: 'action' | 'comedy' | 'documentary' | 'drama' | 'horror' | 'family';
+  title: string;
+}
+
+type ContentProps = {
+  selectedGenreId : number
+}
+export function Content(props : ContentProps) {
+  const [selectedGenre, setSelectedGenre] = useState<GenreResponseProps>({} as GenreResponseProps);
+  const [movies, setMovies] = useState<MovieProps[]>([]);
+  
+  useEffect(() => {
+    api.get<MovieProps[]>(`movies/?Genre_id=${props.selectedGenreId}`).then(response => {
+      setMovies(response.data);
+    });
+    api.get<GenreResponseProps>(`genres/${props.selectedGenreId}`).then(response => {
+      setSelectedGenre(response.data);
+      console.log(selectedGenre)
+    })
+  }, [props.selectedGenreId]);
+
+  return (
+    <div className="container">
+      <header>
+        <span className="category">Categoria:<span> {selectedGenre.title}</span></span>
+      </header>
+      <main>
+        <div className="movies-list">
+          {movies.map(movie => (
+            <MovieCard key ={movie.imdbID} title={movie.Title} poster={movie.Poster} runtime={movie.Runtime} rating={movie.Ratings[0].Value} />
+          ))}
+        </div>
+      </main>
+    </div>
+  )
 }
