+++ src/components/SideBar.tsx
@@ -1,3 +1,15 @@
-export function SideBar() {
-  // Complete aqui
+import { Children, ReactNode } from "react";
+
+type SideBarProps = {
+  children: ReactNode
+}
+export function SideBar(props : SideBarProps) {
+  return(
+    <nav className="sidebar">
+        <span>Watch<p>Me</p></span>
+        <div className="buttons-container">
+          {props.children}          
+        </div>
+      </nav>
+  )
 }
